Fix completeState Unintentional Reset
When the user rotated the screen when TaskDetailFragment's completeTask() was called, completeState would reset despite the Task already being removed from the database. This was unintentional behavior.

The bug has been removed upon moving the completeState variable to the ViewModel, and scoping the ViewModel to the fragment.

However, more work needs to be done to make this stable.