Separate TaskListAddModifyItem Into Two Fragments
TaskListAddModifyItem was a Fragment that had two separate states, "Add" and "Edit." The presence of a navigation argument "task_id" determined if it would go into Add or Edit state.

Now, it is split into two Fragments, "TaskAddFragment" and "TaskEditFragment". They both still use the same layout, but the underlying code is more understandable, modifiable, and less prone to error.