Change Observer to Single Call in TaskDetail
Before, we would set an observer in order to bind the Views to retrieveTask(id)'s data. However, it did not need to be observed, since the data should never change.

Now, bind() in TaskDetailFragment handles the single call to the views in a coroutine, removing the need for LiveData.