Fix Irregular Dragging in TaskListFragment
When TaskListFragment wasn't loaded immediately in HomeFragment's ViewPager2, ViewHolders could be picked up and dragged but they would never swap or even attempt to swap with the other ViewHolders (canDropOver() was never called in the ItemTouchHelper).

This solution also provides a better user experience, since both fragments in the ViewPager2 will be loaded at once. It was a bit jarring for there to be buffer when first opening the app, since there is only two pages.